refactor: cleanup util package
