+++ pkg/postage/listener/listener.go
@@ -20,7 +20,7 @@ import (
 	"github.com/ethersphere/bee/pkg/postage"
 	"github.com/ethersphere/bee/pkg/postage/batchservice"
 	"github.com/ethersphere/bee/pkg/transaction"
-	"github.com/ethersphere/bee/pkg/util"
+	"github.com/ethersphere/bee/pkg/util/syncutil"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
@@ -59,7 +59,7 @@ type listener struct {
 	metrics                     metrics
 	stallingTimeout             time.Duration
 	backoffTime                 time.Duration
-	syncingStopped              *util.Signaler
+	syncingStopped              *syncutil.Signaler
 
 	// Cached postage stamp contract event topics.
 	batchCreatedTopic       common.Hash
@@ -69,7 +69,7 @@ type listener struct {
 }
 
 func New(
-	syncingStopped *util.Signaler,
+	syncingStopped *syncutil.Signaler,
 	logger log.Logger,
 	ev BlockHeightContractFilterer,
 	postageStampContractAddress common.Address,
