+++ pkg/storer/storer.go
@@ -37,7 +37,6 @@ import (
 	localmigration "github.com/ethersphere/bee/pkg/storer/migration"
 	"github.com/ethersphere/bee/pkg/swarm"
 	"github.com/ethersphere/bee/pkg/topology"
-	"github.com/ethersphere/bee/pkg/util"
 	"github.com/ethersphere/bee/pkg/util/syncutil"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/spf13/afero"
@@ -492,7 +491,7 @@ type DB struct {
 	directUploadLimiter chan struct{}
 
 	reserve          *reserve.Reserve
-	inFlight         *util.WaitingCounter
+	inFlight         sync.WaitGroup
 	reserveBinEvents *events.Subscriber
 	baseAddr         swarm.Address
 	batchstore       postage.Storer
@@ -599,7 +598,6 @@ func New(ctx context.Context, dirPath string, opts *Options) (*DB, error) {
 			wakeupDuration: opts.ReserveWakeUpDuration,
 		},
 		directUploadLimiter: make(chan struct{}, pusher.ConcurrentPushes),
-		inFlight:            new(util.WaitingCounter),
 	}
 
 	if db.validStamp == nil {
@@ -660,7 +658,7 @@ func (db *DB) Close() error {
 	bgReserveWorkersClosed := make(chan struct{})
 	go func() {
 		defer close(bgReserveWorkersClosed)
-		if c := db.inFlight.Wait(5 * time.Second); c > 0 {
+		if !syncutil.WaitWithTimeout(&db.inFlight, 2*time.Second) {
 			db.logger.Warning("db shutting down with running goroutines")
 		}
 	}()
