+++ pkg/util/doc.go
@@ -1,0 +1,9 @@
+// Copyright 2023 The Swarm Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+// Package util is a placeholder for common utilities used by multiple packages.
+// It is not intended to be used by other packages and therefore should not be
+// imported or contain any functions, constants, or types. Packages in this
+// directory should have the suffix "util" in their name.
+package util
